Remove full mode; simplify spec_helper
