+++ Gemfile
@@ -13,4 +13,5 @@ end
 group :test do
   gem 'webmock', require: false
   gem 'vcr', require: false
+  gem 'codecov', :require => false
 end
