+++ build_travis.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 if [ "$TRAVIS_SECURE_ENV_VARS" == "true" ]; then 
-  MODE=full bundle exec rake all
+  bundle exec rake all
 else 
   bundle exec rake spec
 fi
