--- spec/spec_helper_full.rb
@@ -1,24 +1,0 @@
-require 'coveralls'
-Coveralls.wear!
-
-require 'simplecov'
-SimpleCov.start do
-  add_filter '/spec/'
-  add_filter '/spec/integration'
-end
-
-require 'pry'
-
-RSpec.configure do |config|
-  config.filter_run focus: true
-  config.run_all_when_everything_filtered = true
-  config.include Credentials
-end
-
-def wait(time, increment = 5, elapsed_time = 0, &block)
-  yield
-rescue RSpec::Expectations::ExpectationNotMetError => e
-  raise e if elapsed_time >= time
-  sleep increment
-  wait(time, increment, elapsed_time + increment, &block)
-end
