--- spec/spec_helper_unit.rb
@@ -1,3 +1,0 @@
-RSpec.configure do |config|
-  config.fail_fast = true
-end
